+++ vk-bot-main/src/main/java/com/barinthecityshow/bot/state/ConcurrentMapState.java
@@ -6,24 +6,24 @@ import com.barinthecityshow.bot.dialog.QuestionAnswer;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 
-public enum ConcurrentMapState implements State<Long, ChainElement<QuestionAnswer>> {
+public enum ConcurrentMapState implements State<Integer, ChainElement<QuestionAnswer>> {
     INSTANCE;
 
-    private Map<Long, ChainElement<QuestionAnswer>> state = new ConcurrentHashMap<>();
+    private Map<Integer, ChainElement<QuestionAnswer>> state = new ConcurrentHashMap<>();
 
 
     @Override
-    public void put(Long key, ChainElement<QuestionAnswer> value) {
+    public void put(Integer key, ChainElement<QuestionAnswer> value) {
         state.put(key, value);
     }
 
     @Override
-    public ChainElement<QuestionAnswer> get(Long key) {
+    public ChainElement<QuestionAnswer> get(Integer key) {
         return state.get(key);
     }
 
     @Override
-    public boolean containsKey(Long key) {
+    public boolean containsKey(Integer key) {
         return state.containsKey(key);
     }
 }
